# swagger/swagger.json
{"swagger":"2.0","info":{"title":"rucci.io","description":"Blog for Nick Rucci","contact":{"name":"Nick Rucci","email":"nick@rucci.io","url":"http://rucci.io"},"license":{"name":"MIT","url":"https://github.com/nicholasrucci/goa-atlas/blob/master/LICENSE"},"version":""},"host":"localhost","basePath":"/","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"auth/refresh":{"post":{"description":"Obtain a refreshed access token","operationId":"auth#refresh","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/RefreshAuthPayload"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Authorize"}}},"schemes":["http"]}},"auth/token":{"post":{"description":"Obtain an access token","operationId":"auth#token","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/TokenAuthPayload"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Authorize"}}},"schemes":["http"]}},"auth/{provider}":{"get":{"description":"OAUTH2 login endpoint","operationId":"auth#oauth","parameters":[{"name":"provider","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Authorize"}}},"schemes":["http"]}},"auth/{provider}/callback":{"get":{"description":"OAUTH2 callback endpoint","operationId":"auth#callback","parameters":[{"name":"provider","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK"}},"schemes":["http"]}},"posts":{"post":{"description":"Create new post","operationId":"post#create","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/CreatePostPayload"}}],"responses":{"201":{"description":"Resource created","headers":{"Location":{"description":"href to created resource","type":"string","pattern":"/posts/[0-9]+"}}}},"schemes":["http"]}},"posts/{postID}":{"get":{"description":"Retrieve post with given id.","operationId":"post#show","parameters":[{"name":"postID","in":"path","description":"Post ID","required":true,"type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"404":{"description":"Not Found"}},"schemes":["http"]},"put":{"description":"Change post attributes","operationId":"post#update","parameters":[{"name":"postID","in":"path","description":"Post ID","required":true,"type":"integer"},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/UpdatePostPayload"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}},"schemes":["http"]},"delete":{"operationId":"post#delete","parameters":[{"name":"postID","in":"path","description":"Post ID","required":true,"type":"integer"}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}},"schemes":["http"]}},"users":{"post":{"description":"Create new user","operationId":"user#create","parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/CreateUserPayload"}}],"responses":{"201":{"description":"Resource created","headers":{"Location":{"description":"href to created resource","type":"string","pattern":"/user/[0-9]+"}}}},"schemes":["http"]}},"users/{userID}":{"get":{"description":"Retrieve user with given id.","operationId":"user#show","parameters":[{"name":"userID","in":"path","description":"User ID","required":true,"type":"integer"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/User"}},"404":{"description":"Not Found"}},"schemes":["http"]},"put":{"description":"Change user attributes","operationId":"user#update","parameters":[{"name":"userID","in":"path","description":"User ID","required":true,"type":"integer"},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/UpdateUserPayload"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}},"schemes":["http"]},"delete":{"operationId":"user#delete","parameters":[{"name":"userID","in":"path","description":"User ID","required":true,"type":"integer"}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}},"schemes":["http"]}}},"definitions":{"Authorize":{"title":"Mediatype identifier: application/vnd.authorize+json","type":"object","properties":{"access_token":{"type":"string","description":"access token","example":"Et quis."},"expires_in":{"type":"integer","description":"Time to expiration in seconds","example":3464305043170853352,"format":"int64"},"token_type":{"type":"string","description":"type of token","example":"Et magnam blanditiis repellat quia quo iure."}},"description":"Token authorization response","example":{"access_token":"Et consectetur non qui reiciendis voluptatem.","expires_in":6018909648102701471,"token_type":"Quam consequatur et ut magnam velit."}},"CreatePostPayload":{"title":"CreatePostPayload","type":"object","properties":{"content":{"type":"string"},"title":{"type":"string"}},"required":["title","content"]},"CreateUserPayload":{"title":"CreateUserPayload","type":"object","properties":{"disabled":{"type":"boolean","description":"Flag for if the user is disabled or not"},"email":{"type":"string","description":"Email of user","format":"email"},"givenName":{"type":"string"},"password":{"type":"string"},"surname":{"type":"string","description":"Surname of user"}},"required":["email","password"]},"RefreshAuthPayload":{"title":"RefreshAuthPayload","type":"object","properties":{"application":{"type":"string","description":"UUID of requesting application"},"email":{"type":"string","description":"email"},"password":{"type":"string","description":"password"}}},"TokenAuthPayload":{"title":"TokenAuthPayload","type":"object","properties":{"application":{"type":"string","description":"UUID of requesting application"},"email":{"type":"string","description":"email"},"password":{"type":"string","description":"password"}}},"UpdatePostPayload":{"title":"UpdatePostPayload","type":"object","properties":{"content":{"type":"string"},"title":{"type":"string"}}},"UpdateUserPayload":{"title":"UpdateUserPayload","type":"object","properties":{"disabled":{"type":"boolean","description":"Flag for if the user is disabled or not"},"email":{"type":"string","description":"Email of user","format":"email"},"givenName":{"type":"string"},"newPassword":{"type":"string"},"oldPassword":{"type":"string"},"surname":{"type":"string","description":"Surname of user"}}},"User":{"title":"Mediatype identifier: application/vnd.user+json","type":"object","properties":{"disabled":{"type":"boolean","description":"Flag for if the user is disabled or not","example":true},"email":{"type":"string","description":"Email of user","example":"theron_maggio@ohararuecker.org","format":"email"},"given_name":{"type":"string","description":"Given name of user","example":"Doloremque sint ipsam aut rerum voluptas."},"href":{"type":"string","description":"API href of the user","example":"Ab amet occaecati ut voluptates."},"id":{"type":"integer","description":"ID of user","example":1374997149414959730,"format":"int64"},"passwordHash":{"type":"string","description":"Hashed password of user","example":"Optio accusantium qui."},"passwordSalt":{"type":"string","description":"Salted password of user","example":"Voluptatem odit nemo commodi asperiores non."},"surname":{"type":"string","description":"Surname of user","example":"Repellendus rem possimus soluta voluptatem odit sint."}},"description":"A user account","example":{"disabled":false,"email":"Voluptates ducimus.","given_name":"Tempora quis ducimus consequatur excepturi vel et.","href":"Est voluptas aliquid dolorem.","id":1193374830158546343,"passwordHash":"Occaecati molestias facere molestiae quia.","passwordSalt":"Sed excepturi ducimus aut tempore deleniti.","surname":"Minima enim inventore molestias quia iusto."},"required":["id","href","email","passwordSalt","passwordHash","disabled"]}},"responses":{"NoContent":{"description":"No Content"},"NotFound":{"description":"Not Found"},"OK":{"description":"OK","schema":{"$ref":"#/definitions/Authorize"}}}}